Rename test/lib to test/test_lib.
