+++ test/test_lib/event_decoder.js
@@ -1,5 +1,5 @@
 
-const web3 = require('../lib/web3.js');
+const web3 = require('../test_lib/web3.js');
 
 class Event {
     static decodeTransactionResponse(transactionResponse) {
